+++ src/main/java/com/asu/secureBankApp/Response/LoginResponse.java
@@ -2,15 +2,15 @@ package com.asu.secureBankApp.Response;
 
 public class LoginResponse {
 
-	Long userId;
+	int userId;
 	
 	Boolean isSuccess;
 	
-	public Long getUserId() {
+	public int getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Long userId) {
+	public void setUserId(int userId) {
 		this.userId = userId;
 	}
